Move deleteBucket to S3BucketProvisionService
Updated usage and test
Created S3ServerException
