+++ connector/src/test/java/it/eng/connector/integration/catalog/DatasetAPIIntegrationTest.java
@@ -16,6 +16,7 @@ import it.eng.tools.model.ArtifactType;
 import it.eng.tools.repository.ArtifactRepository;
 import it.eng.tools.response.GenericApiResponse;
 import it.eng.tools.s3.properties.S3Properties;
+import it.eng.tools.s3.service.S3BucketProvisionService;
 import it.eng.tools.s3.service.S3ClientService;
 import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
@@ -65,6 +66,9 @@ public class DatasetAPIIntegrationTest extends BaseIntegrationTest {
     private S3ClientService s3ClientService;
 
     @Autowired
+    private S3BucketProvisionService s3BucketProvisionService;
+
+    @Autowired
     private S3Properties s3Properties;
 
     @InjectWireMock
@@ -75,7 +79,7 @@ public class DatasetAPIIntegrationTest extends BaseIntegrationTest {
         catalogRepository.deleteAll();
         artifactRepository.deleteAll();
         datasetRepository.deleteAll();
-        if (s3ClientService.bucketExists(s3Properties.getBucketName())) {
+        if (s3BucketProvisionService.bucketExists(s3Properties.getBucketName())) {
             List<String> files = s3ClientService.listFiles(s3Properties.getBucketName());
             if (files != null) {
                 for (String file : files) {
