+++ connector/src/test/java/it/eng/connector/integration/datatransfer/DataTransferDownloadIntegrationTest.java
@@ -17,6 +17,7 @@ import it.eng.tools.model.Artifact;
 import it.eng.tools.model.ArtifactType;
 import it.eng.tools.repository.ArtifactRepository;
 import it.eng.tools.s3.properties.S3Properties;
+import it.eng.tools.s3.service.S3BucketProvisionService;
 import it.eng.tools.s3.service.S3ClientService;
 import okhttp3.Credentials;
 import org.apache.commons.lang3.StringUtils;
@@ -60,6 +61,8 @@ public class DataTransferDownloadIntegrationTest extends BaseIntegrationTest {
     @Autowired
     private S3ClientService s3ClientService;
     @Autowired
+    private S3BucketProvisionService s3BucketProvisionService;
+    @Autowired
     private S3Properties s3Properties;
 
     private static final String FILE_NAME = "hello.txt";
@@ -71,7 +74,7 @@ public class DataTransferDownloadIntegrationTest extends BaseIntegrationTest {
         transferProcessRepository.deleteAll();
         agreementRepository.deleteAll();
         policyEnforcementRepository.deleteAll();
-        if (s3ClientService.bucketExists(s3Properties.getBucketName())) {
+        if (s3BucketProvisionService.bucketExists(s3Properties.getBucketName())) {
             List<String> files = s3ClientService.listFiles(s3Properties.getBucketName());
             if (files != null) {
                 for (String file : files) {
