+++ tools/src/test/java/it/eng/tools/s3/service/S3ClientServiceImplTest.java
@@ -81,333 +81,6 @@ public class S3ClientServiceImplTest {
         lenient().when(s3ClientProvider.s3AsyncClient(any(S3ClientRequest.class))).thenReturn(s3AsyncClient);
     }
 
-    /*
-        @Test
-        @DisplayName("Should create bucket when it doesn't exist")
-        void createBucket_WhenBucketDoesNotExist() {
-            // Arrange
-            String bucketName = "test-bucket";
-
-            // Mock bucket check - bucket doesn't exist
-            when(s3Client.headBucket(any(HeadBucketRequest.class)))
-                    .thenThrow(NoSuchBucketException.builder().message("Bucket does not exist").build());
-
-            when(s3Client.createBucket(any(CreateBucketRequest.class)))
-                    .thenReturn(CreateBucketResponse.builder().build());
-
-            // Act
-            assertDoesNotThrow(() -> s3ClientService.createBucket(bucketName));
-
-            // Assert
-            verify(s3Client).headBucket(any(HeadBucketRequest.class));
-            verify(s3Client).createBucket(any(CreateBucketRequest.class));
-        }
-
-        @Test
-        @DisplayName("Should not create bucket when it already exists")
-        void createBucket_WhenBucketExists() {
-            // Arrange
-            String bucketName = "test-bucket";
-
-            // Mock bucket check - bucket exists
-            when(s3Client.headBucket(any(HeadBucketRequest.class)))
-                    .thenReturn(HeadBucketResponse.builder().build());
-
-            // Act
-            assertDoesNotThrow(() -> s3ClientService.createBucket(bucketName));
-
-            // Assert
-            verify(s3Client).headBucket(any(HeadBucketRequest.class));
-            verify(s3Client, never()).createBucket(any(CreateBucketRequest.class));
-        }
-
-        @Test
-        @DisplayName("Should throw RuntimeException when bucket creation fails")
-        void createBucket_WhenCreationFails() {
-            // Arrange
-            String bucketName = "test-bucket";
-
-            // Mock bucket check - bucket doesn't exist
-            when(s3Client.headBucket(any(HeadBucketRequest.class)))
-                    .thenThrow(NoSuchBucketException.builder().message("Bucket does not exist").build());
-
-            // Mock creation failure
-            when(s3Client.createBucket(any(CreateBucketRequest.class)))
-                    .thenThrow(S3Exception.builder().message("Creation failed").build());
-
-            // Act & Assert
-            RuntimeException exception = assertThrows(RuntimeException.class,
-                    () -> s3ClientService.createBucket(bucketName));
-
-            assertTrue(exception.getMessage().contains("Error creating bucket"));
-            verify(s3Client).headBucket(any(HeadBucketRequest.class));
-            verify(s3Client).createBucket(any(CreateBucketRequest.class));
-        }
-
-        @Test
-        @DisplayName("Should throw RuntimeException when bucket existence check fails")
-        void createBucket_WhenExistenceCheckFails() {
-            // Arrange
-            String bucketName = "test-bucket";
-
-            // Mock bucket check failure
-            when(s3Client.headBucket(any(HeadBucketRequest.class)))
-                    .thenThrow(S3Exception.builder().message("Connection failed").build());
-
-            // Act & Assert
-            RuntimeException exception = assertThrows(RuntimeException.class,
-                    () -> s3ClientService.createBucket(bucketName));
-
-            assertTrue(exception.getMessage().contains("Error creating bucket"));
-            verify(s3Client).headBucket(any(HeadBucketRequest.class));
-            verify(s3Client, never()).createBucket(any(CreateBucketRequest.class));
-        }
-
-        @Test
-        @DisplayName("Should throw IllegalArgumentException when bucket name is null")
-        void createBucket_WhenBucketNameIsNull() {
-            // Act & Assert
-            IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                    () -> s3ClientService.createBucket(null));
-
-            assertEquals("Bucket name cannot be null or empty", exception.getMessage());
-            verify(s3Client, never()).headBucket(any(HeadBucketRequest.class));
-            verify(s3Client, never()).createBucket(any(CreateBucketRequest.class));
-        }
-
-        @Test
-        @DisplayName("Should throw IllegalArgumentException when bucket name is empty")
-        void createBucket_WhenBucketNameIsEmpty() {
-            // Act & Assert
-            IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                    () -> s3ClientService.createBucket(""));
-
-            assertEquals("Bucket name cannot be null or empty", exception.getMessage());
-            verify(s3Client, never()).headBucket(any(HeadBucketRequest.class));
-            verify(s3Client, never()).createBucket(any(CreateBucketRequest.class));
-        }
-    */
-    // deleteBucket tests
-    @Test
-    @DisplayName("Should delete bucket when it exists")
-    void deleteBucket_WhenBucketExists() {
-        // Arrange
-        String bucketName = "test-bucket";
-
-        // Mock bucket check - bucket exists
-        when(s3ClientProvider.adminS3Client()).thenReturn(s3Client);
-        when(s3Client.headBucket(any(HeadBucketRequest.class)))
-                .thenReturn(HeadBucketResponse.builder().build());
-
-        when(s3Client.deleteBucket(any(DeleteBucketRequest.class)))
-                .thenReturn(DeleteBucketResponse.builder().build());
-
-        // Act
-        assertDoesNotThrow(() -> s3ClientService.deleteBucket(bucketName));
-
-        // Assert
-        verify(s3Client).headBucket(any(HeadBucketRequest.class));
-        verify(s3Client).deleteBucket(any(DeleteBucketRequest.class));
-    }
-
-    @Test
-    @DisplayName("Should not attempt to delete bucket when it doesn't exist")
-    void deleteBucket_WhenBucketDoesNotExist() {
-        // Arrange
-        String bucketName = "test-bucket";
-        when(s3ClientProvider.adminS3Client()).thenReturn(s3Client);
-        // Mock bucket check - bucket doesn't exist
-        when(s3Client.headBucket(any(HeadBucketRequest.class)))
-                .thenThrow(NoSuchBucketException.builder()
-                        .message("Bucket does not exist")
-                        .build());
-
-        // Act
-        assertDoesNotThrow(() -> s3ClientService.deleteBucket(bucketName));
-
-        // Assert
-        verify(s3Client).headBucket(any(HeadBucketRequest.class));
-        verify(s3Client, never()).deleteBucket(any(DeleteBucketRequest.class));
-    }
-
-    @Test
-    @DisplayName("Should throw RuntimeException when bucket deletion fails")
-    void deleteBucket_WhenDeletionFails() {
-        // Arrange
-        String bucketName = "test-bucket";
-        when(s3ClientProvider.adminS3Client()).thenReturn(s3Client);
-        // Mock bucket check - bucket exists
-        when(s3Client.headBucket(any(HeadBucketRequest.class)))
-                .thenReturn(HeadBucketResponse.builder().build());
-
-        // Mock deletion failure
-        when(s3Client.deleteBucket(any(DeleteBucketRequest.class)))
-                .thenThrow(S3Exception.builder()
-                        .message("Deletion failed")
-                        .build());
-
-        // Act & Assert
-        RuntimeException exception = assertThrows(RuntimeException.class,
-                () -> s3ClientService.deleteBucket(bucketName));
-
-        assertTrue(exception.getMessage().contains("Error deleting bucket"));
-        verify(s3Client).headBucket(any(HeadBucketRequest.class));
-        verify(s3Client).deleteBucket(any(DeleteBucketRequest.class));
-    }
-
-    @Test
-    @DisplayName("Should throw RuntimeException when bucket check fails with unexpected error")
-    void deleteBucket_WhenBucketCheckFails() {
-        // Arrange
-        String bucketName = "test-bucket";
-        when(s3ClientProvider.adminS3Client()).thenReturn(s3Client);
-        // Mock bucket check failure
-        when(s3Client.headBucket(any(HeadBucketRequest.class)))
-                .thenThrow(S3Exception.builder()
-                        .message("Connection failed")
-                        .build());
-
-        // Act & Assert
-        RuntimeException exception = assertThrows(RuntimeException.class,
-                () -> s3ClientService.deleteBucket(bucketName));
-
-        assertTrue(exception.getMessage().contains("Error deleting bucket"));
-        verify(s3Client).headBucket(any(HeadBucketRequest.class));
-        verify(s3Client, never()).deleteBucket(any(DeleteBucketRequest.class));
-    }
-
-    @Test
-    @DisplayName("Should throw IllegalArgumentException when bucket name is null")
-    void deleteBucket_WhenBucketNameIsNull() {
-        // Act & Assert
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> s3ClientService.deleteBucket(null));
-
-        assertEquals("Bucket name cannot be null or empty", exception.getMessage());
-        verify(s3Client, never()).headBucket(any(HeadBucketRequest.class));
-        verify(s3Client, never()).deleteBucket(any(DeleteBucketRequest.class));
-    }
-
-    @Test
-    @DisplayName("Should throw IllegalArgumentException when bucket name is empty")
-    void deleteBucket_WhenBucketNameIsEmpty() {
-        // Act & Assert
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> s3ClientService.deleteBucket(""));
-
-        assertEquals("Bucket name cannot be null or empty", exception.getMessage());
-        verify(s3Client, never()).headBucket(any(HeadBucketRequest.class));
-        verify(s3Client, never()).deleteBucket(any(DeleteBucketRequest.class));
-    }
-
-    @Test
-    @DisplayName("Should throw IllegalArgumentException when bucket name format is invalid")
-    void deleteBucket_WhenBucketNameFormatIsInvalid() {
-        // Arrange
-        String bucketName = "Invalid.Bucket.Name";
-
-        // Act & Assert
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> s3ClientService.deleteBucket(bucketName));
-
-        assertEquals("Invalid bucket name format: " + bucketName, exception.getMessage());
-        verify(s3Client, never()).headBucket(any(HeadBucketRequest.class));
-        verify(s3Client, never()).deleteBucket(any(DeleteBucketRequest.class));
-    }
-
-    //bucketExists test cases
-    @Test
-    @DisplayName("Should return true when bucket exists")
-    void bucketExists_WhenBucketExists() {
-        // Arrange
-        String bucketName = "test-bucket";
-        when(s3ClientProvider.adminS3Client()).thenReturn(s3Client);
-        when(s3Client.headBucket(any(HeadBucketRequest.class)))
-                .thenReturn(HeadBucketResponse.builder().build());
-
-        // Act
-        boolean result = s3ClientService.bucketExists(bucketName);
-
-        // Assert
-        assertTrue(result);
-        verify(s3Client).headBucket(any(HeadBucketRequest.class));
-    }
-
-    @Test
-    @DisplayName("Should return false when bucket does not exist")
-    void bucketExists_WhenBucketDoesNotExist() {
-        // Arrange
-        String bucketName = "test-bucket";
-        when(s3ClientProvider.adminS3Client()).thenReturn(s3Client);
-
-        when(s3Client.headBucket(any(HeadBucketRequest.class)))
-                .thenThrow(NoSuchBucketException.builder()
-                        .message("The specified bucket does not exist")
-                        .build());
-
-        // Act
-        boolean result = s3ClientService.bucketExists(bucketName);
-
-        // Assert
-        assertFalse(result);
-        verify(s3Client).headBucket(any(HeadBucketRequest.class));
-    }
-
-    @Test
-    @DisplayName("Should throw RuntimeException when checking bucket existence fails")
-    void bucketExists_WhenCheckFails() {
-        // Arrange
-        String bucketName = "test-bucket";
-        when(s3ClientProvider.adminS3Client()).thenReturn(s3Client);
-        when(s3Client.headBucket(any(HeadBucketRequest.class)))
-                .thenThrow(S3Exception.builder()
-                        .message("Connection timeout")
-                        .build());
-
-        // Act & Assert
-        RuntimeException exception = assertThrows(RuntimeException.class,
-                () -> s3ClientService.bucketExists(bucketName));
-
-        assertTrue(exception.getMessage().contains("Error checking if bucket exists"));
-        verify(s3Client).headBucket(any(HeadBucketRequest.class));
-    }
-
-    @Test
-    @DisplayName("Should throw IllegalArgumentException when bucket name is null")
-    void bucketExists_WhenBucketNameIsNull() {
-        // Act & Assert
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> s3ClientService.bucketExists(null));
-
-        assertEquals("Bucket name cannot be null or empty", exception.getMessage());
-        verify(s3Client, never()).headBucket(any(HeadBucketRequest.class));
-    }
-
-    @Test
-    @DisplayName("Should throw IllegalArgumentException when bucket name is empty")
-    void bucketExists_WhenBucketNameIsEmpty() {
-        // Act & Assert
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> s3ClientService.bucketExists(""));
-
-        assertEquals("Bucket name cannot be null or empty", exception.getMessage());
-        verify(s3Client, never()).headBucket(any(HeadBucketRequest.class));
-    }
-
-    @Test
-    @DisplayName("Should throw IllegalArgumentException when bucket name format is invalid")
-    void bucketExists_WhenBucketNameFormatIsInvalid() {
-        // Arrange
-        String bucketName = "Invalid.Bucket.Name";
-
-        // Act & Assert
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> s3ClientService.bucketExists(bucketName));
-
-        assertEquals("Invalid bucket name format: " + bucketName, exception.getMessage());
-        verify(s3Client, never()).headBucket(any(HeadBucketRequest.class));
-    }
-
     // uploadFile tests
     @Test
     @DisplayName("Should successfully upload file")
